modify data type and precision

+++ 0x00-math_complex/main.h
@@ -1,6 +1,8 @@
 #ifndef COMPLEX_H
 #define COMPLEX_H
 
+#include <stdio.h>
+
 /**
  * struct complex - store the real and imaginary
  * part of a complex number
@@ -10,8 +12,8 @@
  */
 typedef struct complex
 {
-	int re;
-	int im;
+	double re;
+	double im;
 } complex;
 
 /* functions prototypes */
